feat: Added a correct logic to create ProgressReport for its date selection [#45]
+ Refactored date related utility functions

+ for loop to create IndicatorReport per Reportable

+ for loop to create IndicatorLocationData per IndicatorReport
